move src/lib/io.ts to src/lib/interfaces/io.ts
